requirements.txt
python-telegram-bot
python-telegram-bot[job-queue]
requests
beautifulsoup4
peewee
psycopg2-binary
selenium
bs4
debugpy
pandas
openpyxl